First round of transferring mem management from DM to ML
